feat(companion): Audit Log表示とヘッダー操作を整理
