+++ docs/task-backlog.md
@@ -1,6 +1,6 @@
 # Task Backlog
 
-- 更新日: 2026-04-04
+- 更新日: 2026-04-30
 - 対象: GitHub issue と repo 内の残タスクの統合管理
 
 ## 方針
@@ -24,6 +24,7 @@
 
 | Priority | 実装状況 | Source | ID | テーマ | 概要 | 依存 / メモ |
 | --- | --- | --- | --- | --- | --- | --- |
+| P2 | 未着手 | Local | `audit-log-storage-split` | AuditLog の重い payload 外出し | `raw_items_json` や provider response が巨大化すると SQLite read / Electron IPC / renderer 表示が詰まるため、AuditLog を summary / small detail / blob payload に分割する | 今回の Companion AuditLog 追加とは別タスク。DB には一覧・検索・小さい metadata を残し、raw items / full response / operation detail は file blob 化または section 別 lazy load にする。現行の `audit_logs.session_id` は `sessions(id)` 参照なので、Companion の実保存対応と削除済み session の audit 保持方針も同時に再検討する |
 | P1 | 完了 | GitHub | [#36](https://github.com/natumekazuki/WithMate/issues/36) | メモリー生成の監査ログ保存不具合 | memory generation 自体は completed update 済みで、renderer の `Audit Log` 再読込条件が stale だった問題を修正した | background activity の `status / updatedAt` 変化も再読込条件へ含め、modal を開いたままでも response / error / raw items の確定値が反映される |
 | P1 | 完了 | GitHub | [#35](https://github.com/natumekazuki/WithMate/issues/35) | メモリー生成のタイムアウト短い | `Memory Extraction` と `Character Reflection` に provider ごとの `Timeout Seconds` を追加し、background plane の timeout を settings から調整できるようにした | default は `180s`、normalize は `30..1800s`。Codex は `AbortSignal.timeout`、Copilot は `sendAndWait(timeout)` へ反映する |
 | P1 | 完了 | GitHub | [#40](https://github.com/natumekazuki/WithMate/issues/40) | まだ SessionNotFound が発生する | cached `CopilotSession` の stale recovery で、未確定 step まで partial 扱いになり internal retry が止まっていた問題を修正した | `tool.execution_start` や pending permission は retry blocker から外し、completed / failed / canceled の command だけを `operations` へ残すようにした |
